move matplotlib related tests from test_suf_plotting to test_matplotlib_backend
