+++ nilearn/plotting/surface/tests/test_matplotlib_backend.py
@@ -1,7 +1,20 @@
+# ruff: noqa: ARG001
+
+import re
+import tempfile
+
+import matplotlib.pyplot as plt
 import numpy as np
 import pytest
+from numpy.testing import assert_array_equal
 
 from nilearn.datasets import fetch_surf_fsaverage
+from nilearn.plotting import (
+    plot_surf,
+    plot_surf_contours,
+    plot_surf_roi,
+    plot_surf_stat_map,
+)
 from nilearn.plotting.surface._matplotlib_backend import (
     MATPLOTLIB_VIEWS,
     _compute_facecolors_matplotlib,
@@ -201,4 +214,313 @@ def test_compute_facecolors_matplotlib_deprecation():
             len(mesh.coordinates),
             0.5,
             alpha,
+        )
+
+
+def test_surface_plotting_axes_error(surf_img_1d):
+    """Test error msg for invalid axes."""
+    figure, axes = plt.subplots()
+    with pytest.raises(AttributeError, match="the projection must be '3d'"):
+        plot_surf_stat_map(stat_map=surf_img_1d, axes=axes)
+
+
+def test_plot_surf_with_title(matplotlib_pyplot, in_memory_mesh, bg_map):
+    """Check title in figure."""
+    display = plot_surf(
+        in_memory_mesh, bg_map=bg_map, title="Test title", engine=ENGINE
+    )
+
+    assert len(display.axes) == 1
+    assert display.axes[0].title._text == "Test title"
+
+
+def test_plot_surf_avg_method(matplotlib_pyplot, in_memory_mesh, bg_map):
+    # Plot with avg_method
+    # Test all built-in methods and check
+    faces = in_memory_mesh.faces
+
+    for method in ["mean", "median", "min", "max"]:
+        display = plot_surf(
+            in_memory_mesh,
+            surf_map=bg_map,
+            avg_method=method,
+            engine=ENGINE,
+        )
+        if method == "mean":
+            agg_faces = np.mean(bg_map[faces], axis=1)
+        elif method == "median":
+            agg_faces = np.median(bg_map[faces], axis=1)
+        elif method == "min":
+            agg_faces = np.min(bg_map[faces], axis=1)
+        elif method == "max":
+            agg_faces = np.max(bg_map[faces], axis=1)
+        vmin = np.min(agg_faces)
+        vmax = np.max(agg_faces)
+        agg_faces -= vmin
+        agg_faces /= vmax - vmin
+        cmap = plt.get_cmap(plt.rcParamsDefault["image.cmap"])
+        assert_array_equal(
+            cmap(agg_faces),
+            display._axstack.as_list()[0].collections[0]._facecolors,
+        )
+
+    #  Try custom avg_method
+    def custom_avg_function(vertices):
+        return vertices[0] * vertices[1] * vertices[2]
+
+    plot_surf(
+        in_memory_mesh,
+        surf_map=bg_map,
+        avg_method=custom_avg_function,
+        engine=ENGINE,
+    )
+
+
+def test_plot_surf_avg_method_errors(in_memory_mesh, bg_map):
+    with pytest.raises(
+        ValueError,
+        match=(
+            "Array computed with the custom "
+            "function from avg_method does "
+            "not have the correct shape"
+        ),
+    ):
+
+        def custom_avg_function(vertices):
+            return [vertices[0] * vertices[1], vertices[2]]
+
+        plot_surf(
+            in_memory_mesh,
+            surf_map=bg_map,
+            avg_method=custom_avg_function,
+            engine=ENGINE,
+        )
+
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "avg_method should be either "
+            "['mean', 'median', 'max', 'min'] "
+            "or a custom function"
+        ),
+    ):
+        custom_avg_function = {}
+
+        plot_surf(
+            in_memory_mesh,
+            surf_map=bg_map,
+            avg_method=custom_avg_function,
+            engine=ENGINE,
+        )
+
+        plot_surf(
+            in_memory_mesh,
+            surf_map=bg_map,
+            avg_method="foo",
+            engine=ENGINE,
+        )
+
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Array computed with the custom function "
+            "from avg_method should be an array of "
+            "numbers (int or float)"
+        ),
+    ):
+
+        def custom_avg_function(vertices):
+            return "string"
+
+        plot_surf(
+            in_memory_mesh,
+            surf_map=bg_map,
+            avg_method=custom_avg_function,
+            engine=ENGINE,
+        )
+
+
+def test_plot_surf_stat_map_matplotlib_specific(
+    matplotlib_pyplot, in_memory_mesh, bg_map
+):
+    # Plot to axes
+    axes = plt.subplots(ncols=2, subplot_kw={"projection": "3d"})[1]
+    for ax in axes.flatten():
+        plot_surf_stat_map(in_memory_mesh, stat_map=bg_map, axes=ax)
+    axes = plt.subplots(ncols=2, subplot_kw={"projection": "3d"})[1]
+    for ax in axes.flatten():
+        plot_surf_stat_map(in_memory_mesh, stat_map=bg_map, axes=ax)
+
+    fig = plot_surf_stat_map(in_memory_mesh, stat_map=bg_map, colorbar=False)
+
+    assert len(fig.axes) == 1
+
+    # symmetric_cbar
+    fig = plot_surf_stat_map(
+        in_memory_mesh, stat_map=bg_map, symmetric_cbar=True
+    )
+    fig.canvas.draw()
+
+    assert len(fig.axes) == 2
+
+    yticklabels = fig.axes[1].get_yticklabels()
+    first, last = yticklabels[0].get_text(), yticklabels[-1].get_text()
+
+    assert float(first) == -float(last)
+
+    # no symmetric_cbar
+    fig = plot_surf_stat_map(
+        in_memory_mesh, stat_map=bg_map, symmetric_cbar=False
+    )
+    fig.canvas.draw()
+
+    assert len(fig.axes) == 2
+
+    yticklabels = fig.axes[1].get_yticklabels()
+    first, last = yticklabels[0].get_text(), yticklabels[-1].get_text()
+
+    assert float(first) != -float(last)
+
+    # Test handling of nan values in texture data
+    # Add nan values in the texture
+    bg_map[2] = np.nan
+    # Plot the surface stat map
+    fig = plot_surf_stat_map(in_memory_mesh, stat_map=bg_map)
+    # Check that the resulting plot facecolors contain no transparent faces
+    # (last column equals zero) even though the texture contains nan values
+    tmp = fig._axstack.as_list()[0].collections[0]
+
+    assert (
+        in_memory_mesh.faces.shape[0] == ((tmp._facecolors[:, 3]) != 0).sum()
+    )
+
+
+def test_plot_surf_roi_matplotlib_specific(
+    matplotlib_pyplot, surface_image_roi
+):
+    # change vmin, vmax
+    img = plot_surf_roi(
+        surface_image_roi.mesh,
+        roi_map=surface_image_roi,
+        vmin=1.2,
+        vmax=8.9,
+        colorbar=True,
+        engine="matplotlib",
+    )
+    img.canvas.draw()
+    cbar = img.axes[-1]
+    cbar_vmin = float(cbar.get_yticklabels()[0].get_text())
+    cbar_vmax = float(cbar.get_yticklabels()[-1].get_text())
+
+    assert cbar_vmin == 1.0
+    assert cbar_vmax == 8.0
+
+    img2 = plot_surf_roi(
+        surface_image_roi.mesh,
+        roi_map=surface_image_roi,
+        vmin=1.2,
+        vmax=8.9,
+        colorbar=True,
+        cbar_tick_format="%.2g",
+        engine="matplotlib",
+    )
+    img2.canvas.draw()
+    cbar = img2.axes[-1]
+    cbar_vmin = float(cbar.get_yticklabels()[0].get_text())
+    cbar_vmax = float(cbar.get_yticklabels()[-1].get_text())
+
+    assert cbar_vmin == 1.2
+    assert cbar_vmax == 8.9
+
+
+def test_plot_surf_roi_matplotlib_specific_nan_handling(
+    matplotlib_pyplot,
+    surface_image_parcellation,
+):
+    # Test nans handling
+    surface_image_parcellation.data.parts["left"][::2] = np.nan
+    img = plot_surf_roi(
+        surface_image_parcellation.mesh,
+        roi_map=surface_image_parcellation,
+        engine="matplotlib",
+        hemi="left",
+    )
+    # Check that the resulting plot facecolors contain no transparent faces
+    # (last column equals zero) even though the texture contains nan values
+    tmp = img._axstack.as_list()[0].collections[0]
+    n_faces = surface_image_parcellation.mesh.parts["left"].faces.shape[0]
+
+    assert n_faces == ((tmp._facecolors[:, 3]) != 0).sum()
+
+
+def test_plot_surf_roi_matplotlib_specific_plot_to_axes(
+    matplotlib_pyplot, surface_image_roi
+):
+    """Test plotting directly on some axes."""
+    plot_surf_roi(
+        surface_image_roi.mesh,
+        roi_map=surface_image_roi,
+        axes=None,
+        figure=plt.gcf(),
+        engine=ENGINE,
+    )
+
+    _, ax = plt.subplots(subplot_kw={"projection": "3d"})
+
+    with tempfile.NamedTemporaryFile() as tmp_file:
+        plot_surf_roi(
+            surface_image_roi.mesh,
+            roi_map=surface_image_roi,
+            axes=ax,
+            figure=None,
+            output_file=tmp_file.name,
+            engine=ENGINE,
+        )
+
+    with tempfile.NamedTemporaryFile() as tmp_file:
+        plot_surf_roi(
+            surface_image_roi.mesh,
+            roi_map=surface_image_roi,
+            axes=ax,
+            figure=None,
+            output_file=tmp_file.name,
+            colorbar=True,
+            engine=ENGINE,
+        )
+
+
+def test_plot_surf_contours_fig_axes(
+    matplotlib_pyplot, in_memory_mesh, parcellation
+):
+    fig, axes = plt.subplots(1, 1, subplot_kw={"projection": "3d"})
+    plot_surf_contours(in_memory_mesh, parcellation, axes=axes)
+    plot_surf_contours(in_memory_mesh, parcellation, figure=fig)
+
+
+def test_plot_surf_contours_error(rng, in_memory_mesh, parcellation):
+    # we need an invalid parcellation for testing
+    invalid_parcellation = rng.uniform(size=(in_memory_mesh.n_vertices))
+    with pytest.raises(
+        ValueError, match="Vertices in parcellation do not form region."
+    ):
+        plot_surf_contours(in_memory_mesh, invalid_parcellation)
+
+    _, axes = plt.subplots(1, 1)
+    with pytest.raises(ValueError, match="Axes must be 3D."):
+        plot_surf_contours(in_memory_mesh, parcellation, axes=axes)
+
+    msg = "All elements of colors .* matplotlib .* RGBA"
+    with pytest.raises(ValueError, match=msg):
+        plot_surf_contours(
+            in_memory_mesh, parcellation, levels=[1, 2], colors=[[1, 2], 3]
+        )
+
+    msg = "Levels, labels, and colors argument .* same length or None."
+    with pytest.raises(ValueError, match=msg):
+        plot_surf_contours(
+            in_memory_mesh,
+            parcellation,
+            levels=[1, 2],
+            colors=["r"],
+            labels=["1", "2"],
         )
