+++ nilearn/plotting/surface/tests/test_surf_plotting.py
@@ -2,16 +2,13 @@
 
 # ruff: noqa: ARG001
 
-import re
 import tempfile
 from unittest import mock
 
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import pytest
 from matplotlib.figure import Figure
-from numpy.testing import assert_array_equal
 
 from nilearn._utils.helpers import is_kaleido_installed, is_plotly_installed
 from nilearn.datasets import fetch_surf_fsaverage
@@ -23,7 +20,6 @@ from nilearn.plotting import (
     plot_surf_stat_map,
 )
 from nilearn.plotting.displays import PlotlySurfaceFigure, SurfaceFigure
-from nilearn.surface import SurfaceImage
 
 try:
     import IPython.display  # noqa:F401
@@ -31,12 +27,6 @@ except ImportError:
     IPYTHON_INSTALLED = False
 else:
     IPYTHON_INSTALLED = True
-
-
-@pytest.fixture
-def bg_map(rng, in_memory_mesh):
-    """Return a background map with positive value."""
-    return np.abs(rng.standard_normal(size=in_memory_mesh.n_vertices))
 
 
 @pytest.mark.parametrize(
@@ -64,13 +54,6 @@ def test_check_surface_plotting_inputs_errors():
         plot_surf_contours(roi_map=1, surf_mesh=None)
     with pytest.raises(TypeError, match="must be a SurfaceImage instance"):
         plot_surf_roi(roi_map=1, surf_mesh=None)
-
-
-def test_surface_plotting_axes_error(surf_img_1d):
-    """Test error msg for invalid axes."""
-    figure, axes = plt.subplots()
-    with pytest.raises(AttributeError, match="the projection must be '3d'"):
-        plot_surf_stat_map(stat_map=surf_img_1d, axes=axes)
 
 
 def test_plot_surf_contours_warning_hemi(in_memory_mesh):
@@ -266,22 +249,6 @@ def surf_roi_data(rng, in_memory_mesh):
     return roi_map
 
 
-@pytest.fixture
-def surface_image_roi(surf_mask_1d):
-    """SurfaceImage for plotting."""
-    return surf_mask_1d
-
-
-@pytest.fixture
-def surface_image_parcellation(rng, in_memory_mesh):
-    data = rng.integers(100, size=(in_memory_mesh.n_vertices, 1)).astype(float)
-    parcellation = SurfaceImage(
-        mesh={"left": in_memory_mesh, "right": in_memory_mesh},
-        data={"left": data, "right": data},
-    )
-    return parcellation
-
-
 def test_add_contours(plotly, surface_image_roi):
     """Test that add_contours updates data in PlotlySurfaceFigure."""
     figure = plot_surf(surface_image_roi.mesh, engine="plotly")
@@ -449,58 +416,6 @@ def test_plot_surf_hemi_views_plotly(
     )
 
 
-def test_plot_surf_with_title(matplotlib_pyplot, in_memory_mesh, bg_map):
-    """Check title in figure."""
-    display = plot_surf(
-        in_memory_mesh, bg_map=bg_map, title="Test title", engine="matplotlib"
-    )
-
-    assert len(display.axes) == 1
-    assert display.axes[0].title._text == "Test title"
-
-
-def test_plot_surf_avg_method(matplotlib_pyplot, in_memory_mesh, bg_map):
-    # Plot with avg_method
-    # Test all built-in methods and check
-    faces = in_memory_mesh.faces
-
-    for method in ["mean", "median", "min", "max"]:
-        display = plot_surf(
-            in_memory_mesh,
-            surf_map=bg_map,
-            avg_method=method,
-            engine="matplotlib",
-        )
-        if method == "mean":
-            agg_faces = np.mean(bg_map[faces], axis=1)
-        elif method == "median":
-            agg_faces = np.median(bg_map[faces], axis=1)
-        elif method == "min":
-            agg_faces = np.min(bg_map[faces], axis=1)
-        elif method == "max":
-            agg_faces = np.max(bg_map[faces], axis=1)
-        vmin = np.min(agg_faces)
-        vmax = np.max(agg_faces)
-        agg_faces -= vmin
-        agg_faces /= vmax - vmin
-        cmap = plt.get_cmap(plt.rcParamsDefault["image.cmap"])
-        assert_array_equal(
-            cmap(agg_faces),
-            display._axstack.as_list()[0].collections[0]._facecolors,
-        )
-
-    #  Try custom avg_method
-    def custom_avg_function(vertices):
-        return vertices[0] * vertices[1] * vertices[2]
-
-    plot_surf(
-        in_memory_mesh,
-        surf_map=bg_map,
-        avg_method=custom_avg_function,
-        engine="matplotlib",
-    )
-
-
 @pytest.mark.parametrize("engine", ["matplotlib", "plotly"])
 def test_plot_surf_error(engine, rng, in_memory_mesh):
     if not is_plotly_installed() and engine == "plotly":
@@ -553,70 +468,6 @@ def test_plot_surf_warnings_not_implemented_in_plotly(
             surf_map=bg_map,
             engine="plotly",
             **kwargs,
-        )
-
-
-def test_plot_surf_avg_method_errors(in_memory_mesh, bg_map):
-    with pytest.raises(
-        ValueError,
-        match=(
-            "Array computed with the custom "
-            "function from avg_method does "
-            "not have the correct shape"
-        ),
-    ):
-
-        def custom_avg_function(vertices):
-            return [vertices[0] * vertices[1], vertices[2]]
-
-        plot_surf(
-            in_memory_mesh,
-            surf_map=bg_map,
-            avg_method=custom_avg_function,
-            engine="matplotlib",
-        )
-
-    with pytest.raises(
-        ValueError,
-        match=re.escape(
-            "avg_method should be either "
-            "['mean', 'median', 'max', 'min'] "
-            "or a custom function"
-        ),
-    ):
-        custom_avg_function = {}
-
-        plot_surf(
-            in_memory_mesh,
-            surf_map=bg_map,
-            avg_method=custom_avg_function,
-            engine="matplotlib",
-        )
-
-        plot_surf(
-            in_memory_mesh,
-            surf_map=bg_map,
-            avg_method="foo",
-            engine="matplotlib",
-        )
-
-    with pytest.raises(
-        ValueError,
-        match=re.escape(
-            "Array computed with the custom function "
-            "from avg_method should be an array of "
-            "numbers (int or float)"
-        ),
-    ):
-
-        def custom_avg_function(vertices):
-            return "string"
-
-        plot_surf(
-            in_memory_mesh,
-            surf_map=bg_map,
-            avg_method=custom_avg_function,
-            engine="matplotlib",
         )
 
 
@@ -719,61 +570,6 @@ def test_plot_surf_stat_map_colormap(
     )
 
 
-def test_plot_surf_stat_map_matplotlib_specific(
-    matplotlib_pyplot, in_memory_mesh, bg_map
-):
-    # Plot to axes
-    axes = plt.subplots(ncols=2, subplot_kw={"projection": "3d"})[1]
-    for ax in axes.flatten():
-        plot_surf_stat_map(in_memory_mesh, stat_map=bg_map, axes=ax)
-    axes = plt.subplots(ncols=2, subplot_kw={"projection": "3d"})[1]
-    for ax in axes.flatten():
-        plot_surf_stat_map(in_memory_mesh, stat_map=bg_map, axes=ax)
-
-    fig = plot_surf_stat_map(in_memory_mesh, stat_map=bg_map, colorbar=False)
-
-    assert len(fig.axes) == 1
-
-    # symmetric_cbar
-    fig = plot_surf_stat_map(
-        in_memory_mesh, stat_map=bg_map, symmetric_cbar=True
-    )
-    fig.canvas.draw()
-
-    assert len(fig.axes) == 2
-
-    yticklabels = fig.axes[1].get_yticklabels()
-    first, last = yticklabels[0].get_text(), yticklabels[-1].get_text()
-
-    assert float(first) == -float(last)
-
-    # no symmetric_cbar
-    fig = plot_surf_stat_map(
-        in_memory_mesh, stat_map=bg_map, symmetric_cbar=False
-    )
-    fig.canvas.draw()
-
-    assert len(fig.axes) == 2
-
-    yticklabels = fig.axes[1].get_yticklabels()
-    first, last = yticklabels[0].get_text(), yticklabels[-1].get_text()
-
-    assert float(first) != -float(last)
-
-    # Test handling of nan values in texture data
-    # Add nan values in the texture
-    bg_map[2] = np.nan
-    # Plot the surface stat map
-    fig = plot_surf_stat_map(in_memory_mesh, stat_map=bg_map)
-    # Check that the resulting plot facecolors contain no transparent faces
-    # (last column equals zero) even though the texture contains nan values
-    tmp = fig._axstack.as_list()[0].collections[0]
-
-    assert (
-        in_memory_mesh.faces.shape[0] == ((tmp._facecolors[:, 3]) != 0).sum()
-    )
-
-
 def test_plot_surf_stat_map_error(in_memory_mesh, bg_map):
     # Wrong size of stat map data
     with pytest.raises(
@@ -835,100 +631,6 @@ def test_plot_surf_parcellation_plotly(
         colorbar=colorbar,
         cbar_tick_format=cbar_tick_format,
     )
-
-
-def test_plot_surf_roi_matplotlib_specific(
-    matplotlib_pyplot, surface_image_roi
-):
-    # change vmin, vmax
-    img = plot_surf_roi(
-        surface_image_roi.mesh,
-        roi_map=surface_image_roi,
-        vmin=1.2,
-        vmax=8.9,
-        colorbar=True,
-        engine="matplotlib",
-    )
-    img.canvas.draw()
-    cbar = img.axes[-1]
-    cbar_vmin = float(cbar.get_yticklabels()[0].get_text())
-    cbar_vmax = float(cbar.get_yticklabels()[-1].get_text())
-
-    assert cbar_vmin == 1.0
-    assert cbar_vmax == 8.0
-
-    img2 = plot_surf_roi(
-        surface_image_roi.mesh,
-        roi_map=surface_image_roi,
-        vmin=1.2,
-        vmax=8.9,
-        colorbar=True,
-        cbar_tick_format="%.2g",
-        engine="matplotlib",
-    )
-    img2.canvas.draw()
-    cbar = img2.axes[-1]
-    cbar_vmin = float(cbar.get_yticklabels()[0].get_text())
-    cbar_vmax = float(cbar.get_yticklabels()[-1].get_text())
-
-    assert cbar_vmin == 1.2
-    assert cbar_vmax == 8.9
-
-
-def test_plot_surf_roi_matplotlib_specific_nan_handling(
-    matplotlib_pyplot,
-    surface_image_parcellation,
-):
-    # Test nans handling
-    surface_image_parcellation.data.parts["left"][::2] = np.nan
-    img = plot_surf_roi(
-        surface_image_parcellation.mesh,
-        roi_map=surface_image_parcellation,
-        engine="matplotlib",
-        hemi="left",
-    )
-    # Check that the resulting plot facecolors contain no transparent faces
-    # (last column equals zero) even though the texture contains nan values
-    tmp = img._axstack.as_list()[0].collections[0]
-    n_faces = surface_image_parcellation.mesh.parts["left"].faces.shape[0]
-
-    assert n_faces == ((tmp._facecolors[:, 3]) != 0).sum()
-
-
-def test_plot_surf_roi_matplotlib_specific_plot_to_axes(
-    matplotlib_pyplot, surface_image_roi
-):
-    """Test plotting directly on some axes."""
-    plot_surf_roi(
-        surface_image_roi.mesh,
-        roi_map=surface_image_roi,
-        axes=None,
-        figure=plt.gcf(),
-        engine="matplotlib",
-    )
-
-    _, ax = plt.subplots(subplot_kw={"projection": "3d"})
-
-    with tempfile.NamedTemporaryFile() as tmp_file:
-        plot_surf_roi(
-            surface_image_roi.mesh,
-            roi_map=surface_image_roi,
-            axes=ax,
-            figure=None,
-            output_file=tmp_file.name,
-            engine="matplotlib",
-        )
-
-    with tempfile.NamedTemporaryFile() as tmp_file:
-        plot_surf_roi(
-            surface_image_roi.mesh,
-            roi_map=surface_image_roi,
-            axes=ax,
-            figure=None,
-            output_file=tmp_file.name,
-            colorbar=True,
-            engine="matplotlib",
-        )
 
 
 @pytest.mark.parametrize("engine", ["matplotlib", "plotly"])
@@ -1176,14 +878,6 @@ def test_plot_img_on_surf_input_as_file(matplotlib_pyplot, img_3d_mni_as_file):
     plot_img_on_surf(stat_map=str(img_3d_mni_as_file))
 
 
-@pytest.fixture
-def parcellation(in_memory_mesh):
-    parcellation = np.zeros((in_memory_mesh.n_vertices,))
-    parcellation[in_memory_mesh.faces[3]] = 1
-    parcellation[in_memory_mesh.faces[5]] = 2
-    return parcellation
-
-
 def test_plot_surf_contours(
     matplotlib_pyplot, in_memory_mesh, parcellation, surf_mask_1d
 ):
@@ -1233,14 +927,6 @@ def test_plot_surf_contours_colors(
     )
 
 
-def test_plot_surf_contours_fig_axes(
-    matplotlib_pyplot, in_memory_mesh, parcellation
-):
-    fig, axes = plt.subplots(1, 1, subplot_kw={"projection": "3d"})
-    plot_surf_contours(in_memory_mesh, parcellation, axes=axes)
-    plot_surf_contours(in_memory_mesh, parcellation, figure=fig)
-
-
 def test_plot_surf_contours_axis_title(
     matplotlib_pyplot, in_memory_mesh, parcellation
 ):
@@ -1276,35 +962,6 @@ def test_plot_surf_contours_axis_title(
     with tempfile.NamedTemporaryFile() as tmp_file:
         plot_surf_contours(
             in_memory_mesh, parcellation, output_file=tmp_file.name
-        )
-
-
-def test_plot_surf_contours_error(rng, in_memory_mesh, parcellation):
-    # we need an invalid parcellation for testing
-    invalid_parcellation = rng.uniform(size=(in_memory_mesh.n_vertices))
-    with pytest.raises(
-        ValueError, match="Vertices in parcellation do not form region."
-    ):
-        plot_surf_contours(in_memory_mesh, invalid_parcellation)
-
-    _, axes = plt.subplots(1, 1)
-    with pytest.raises(ValueError, match="Axes must be 3D."):
-        plot_surf_contours(in_memory_mesh, parcellation, axes=axes)
-
-    msg = "All elements of colors .* matplotlib .* RGBA"
-    with pytest.raises(ValueError, match=msg):
-        plot_surf_contours(
-            in_memory_mesh, parcellation, levels=[1, 2], colors=[[1, 2], 3]
-        )
-
-    msg = "Levels, labels, and colors argument .* same length or None."
-    with pytest.raises(ValueError, match=msg):
-        plot_surf_contours(
-            in_memory_mesh,
-            parcellation,
-            levels=[1, 2],
-            colors=["r"],
-            labels=["1", "2"],
         )
